Add files via upload
I made this in like 5 minutes with only a minor inconvenience of my entire computer crashing because I forgot to add a Sleep() function but that doesn't matter. This program just deletes a random amount of characters (1-20) from your sentence every 1 - 5 seconds. This is a good way to type faster I guess.